requirements.txt
torch
diffusers
accelerate
transformers
opencv-python
Pillow
numpy
peft
streamlit
